Fixes after teacher's comments
Fixes in the Token handling and message encrypt and decrypt are made after teacher's comments on the code. The backend part seems to be working now.
The logged in user can read their messages, can create new ones. The messages are enrypted in the database but read in plain text. Next step is to create the client part.

Currently all users use the same hardcoded encryption and decryption key.

+++ timelord/src/main/java/com/tasktwo/timelord/server/DTO/MessageDTO.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 @Getter
 @Setter
 public class MessageDTO {
-    private Long id;
+    private int id;
     private String message;
-    private Long idUser;
+    private int idUser;
 }
